# topgun-services/docs/swagger.yaml
definitions:
...
  models.User:
    properties:
      email:
        type: string
      first_name:
        type: string
      first_name_en:
        type: string
      last_name:
        type: string
      last_name_en:
        type: string
      password:
        type: string
      title:
        type: string
      title_en:
        type: string
    type: object
...
info:
  contact: {}
  title: KKU GS ADMISSION SERVICES API
paths:
...
  /api/v1/users/:
    get:
      consumes:
      - application/json
      description: Get users
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 10
        description: Per Page
        in: query
        name: per_page
        type: integer
      - description: Keyword
        in: query
        name: keyword
        type: string
      - default: '"first_name,last_name"'
        description: Column
        in: query
        name: column
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: GetUsers
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Create user
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: CreateUser
      tags:
      - User
  /api/v1/users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: DeleteUser
      tags:
      - User
    get:
      consumes:
      - application/json
      description: Get user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: GetUser
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: UpdateUser
      tags:
      - User
  /api/v1/users/me:
    get:
      consumes:
      - application/json
      description: Get me
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: GetMe
      tags:
      - User
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"